feat: 마이페이지

+++ src/main/java/com/signalmatch_backend/user/dto/InvestorMyPageResponse.java
@@ -4,9 +4,9 @@ import com.signalmatch_backend.investor.dto.InvestorProfileInfo;
 
 public record InvestorMyPageResponse(
         InvestorProfileInfo profile,
-        int matchedStartupCount
+        long matchedStartupCount
 ) implements MyPageResponse{
-    public static InvestorMyPageResponse of(InvestorProfileInfo profile, int matchedCount) {
+    public static InvestorMyPageResponse of(InvestorProfileInfo profile, long matchedCount) {
         return new InvestorMyPageResponse(profile, matchedCount);
     }
